refactor: wrap lib code in a "src" folder
