--- lib/index.ts
@@ -1,3 +1,0 @@
-export * from './utils/types'
-export * from './utils/validation'
-export { colorYiq as default } from './color-yiq'
